requirements.txt
adjustText==1.3.0
aiobotocore==3.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
anndata==0.12.10
annotated-doc==0.0.4
annotated-types==0.7.0
annsel==0.1.2
anyio==4.12.1
array-api-compat==1.14.0
attrs==25.4.0
beautifulsoup4==4.14.3
botocore==1.42.61
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.3
csbdeep==0.8.2
cycler==0.12.1
dask==2026.1.1
dask-image==2025.11.0
datashader==0.18.2
Deprecated==1.3.1
distributed==2026.1.1
docrep==0.3.2
donfig==0.8.1.post1
enrichmap==0.1.29
esda==2.8.2
fast-array-utils==1.3.1
filelock==3.25.2
FlowIO==1.4.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
geopandas==1.1.3
giddy==2.3.8
google-crc32c==1.8.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
igraph==1.0.0
imagecodecs==2025.11.11
ImageIO==2.37.3
importlib_metadata==8.7.1
inflect==7.5.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
legacy-api-wrap==1.5
libpysal==4.14.1
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.46.0
locket==1.0.0
loguru==0.7.3
mapclassify==2.10.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-scalebar==0.9.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multipledispatch==1.0.0
multiscale_spatial_image==2.0.3
narwhals==2.18.0
natsort==8.4.0
networkx==3.6.1
novae==1.0.3
numba==0.64.0
numcodecs==0.16.5
numpy==2.4.3
ome-types==0.6.3
ome-zarr==0.14.0
omnipath==1.0.12
packaging==26.0
pandas==2.3.3
param==2.3.2
partd==1.4.2
pathlib_abc==0.5.2
patsy==1.0.2
pillow==12.1.1
PIMS==0.7
platformdirs==4.9.4
pooch==1.9.0
progressbar2==4.5.0
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
pyct==0.6.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic_core==2.41.5
pygam==0.12.0
Pygments==2.19.2
pynndescent==0.6.0
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytorch-lightning==2.6.1
pytz==2026.1.post1
PyYAML==6.0.3
quantecon==0.11.1
rangehttpserver==1.4.0
readfcs==2.1.0
requests==2.32.5
rich==14.3.3
s3fs==2026.2.0
safetensors==0.7.0
scanpy==1.11.5
scikit-gstat==1.0.23
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
session-info2==0.4
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
slicerator==1.1.0
sopa==2.2.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spatial_image==1.2.3
spatialdata==0.7.2
spatialdata-io==0.6.0
spatialdata-plot==0.2.14
splot==1.1.7
spreg==1.8.5
squidpy==1.8.1
stardist==0.9.2
statannotations==0.7.2
statsmodels==0.14.6
sympy==1.14.0
tblib==3.2.2
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2026.3.3
toolz==1.1.0
torch==2.10.0
torch-geometric==2.7.0
torchmetrics==1.9.0
tornado==6.5.5
tqdm==4.67.3
typeguard==4.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
universal_pathlib==0.3.10
urllib3==2.6.3
validators==0.35.0
win32_setctime==1.2.0
wrapt==2.1.2
xarray==2026.2.0
xarray-dataclass==3.0.0
xarray-spatial==0.5.3
xmltodict==1.0.4
xsdata==26.2
xxhash==3.6.0
yarl==1.23.0
zarr==3.1.5
zict==3.0.0
zipp==3.23.0
